Adds the option to exclude several folders at once
This commit implements the feature to exclude as many folders as
you want from the cloning process. It also removes these folders
from the size check in order to accurately verify if there is
enough space on the target partition to hold the source partition
data without the excluded folders and doig so resolve the issue
reported in thiggy01#1 by @andbad.